requirements.txt
Flask
requests
pandas
selenium
google-generativeai 